Device editor + IO Mapping table view — full end-to-end IDE
The center pane now switches on a `view` state (app | device | iomap):
- DevicePane shows a Modbus form (host/port/slave/poll + channels CRUD
  table) when the user clicks a device in the project tree. EtherCAT
  selects show a placeholder; ironplc EtherCAT support comes later.
- IoMapPane shows a table of mappings (Application × Variable × Direction
  × Device × Channel) with inline edit, add, and delete. Saves via PUT
  /api/iomap.
- ProjectTree gained a top-level "IO Mapping" entry under Devices for
  navigating to that view, and device rows are now clickable (open) +
  right-click delete. Active selection is highlighted regardless of kind.

Plumbing changes:
- Renamed `RuntimeState.selection` semantics: added `view`,
  `currentDevice`, `iomap`. New actions: `selectDevice`, `openIoMap`,
  `saveDevice`, `saveIomap`. selectApp now also sets view='app'.
- ModbusConfig.poll_interval_ms switched u64 → u32 so it round-trips
  through JSON as a plain number rather than a TS bigint that
  JSON.stringify can't serialize.

Verified end-to-end: edited the demo project's `tank` device to point at
127.0.0.1:5502 with a `coil_0` channel; PUT iomap blink → coil_0
(output); POST /api/run app=main. /api/_demo/slave shows coil 0 toggling
as `blink` flips. Same flow now drivable purely from the browser UI.

+++ crates/project/src/types.rs
@@ -72,7 +72,9 @@ pub struct ModbusConfig {
     pub host: String,
     pub port: u16,
     pub slave_id: u8,
-    pub poll_interval_ms: u64,
+    /// Polling interval in milliseconds (u32 so it round-trips through JSON
+    /// as a number rather than a TS bigint).
+    pub poll_interval_ms: u32,
     #[serde(default)]
     pub channels: Vec<ModbusChannel>,
 }
